arduino/requirements.txt
pyserial==3.5
requests==2.31.0
SQLAlchemy==2.0.29
alembic==1.13.1